Replaced config store and action with reaction
