+++ app/scripts/mainApp.jsx
@@ -3,8 +3,10 @@ import React from 'react';
 import { Router, Route, IndexRoute }  from 'react-router';
 import LessLoader from 'less-hot';
 
-import ConfigActions from './actions/config';
+import State from './stores/state';
+
 import RouteActions from './actions/route';
+import './reactions/config';
 
 import SquelchView from './components/squelchView';
 import WelcomeView from './components/welcome';
@@ -15,13 +17,15 @@ import ChannelView from './components/channel';
 const lessLoader = new LessLoader();
 document.querySelector('head').appendChild(lessLoader(path.join(__dirname, '../less/app.less')));
 
-ConfigActions.load();
+State.trigger('config:load');
 
 const onUpdate = function() {
     RouteActions.changeRoute(this.state);
 };
 
 // TODO: switch to react-dom
+
+// TODO: wrap in root node that re-renders on State 'update' event
 React.render(
     <Router onUpdate={onUpdate}>
         <Route path="/" component={SquelchView}>
