+++ app/scripts/stores/state.jsx
@@ -2,7 +2,9 @@ import Freezer from 'freezer-js';
 
 const State = new Freezer({
     servers: {},
-    messages: {}
+    messages: {},
+    config: {},
+    configDir: ''
 }, { live: false });
 
 export default State;
